+++ contracts/src/v0.8/l2ep/dev/arbitrum/ArbitrumCrossDomainGovernor.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import "../../interfaces/DelegateForwarderInterface.sol";
+import "../interfaces/DelegateForwarderInterface.sol";
 import "../../../vendor/arb-bridge-eth/v0.8.0-custom/contracts/libraries/AddressAliasHelper.sol";
 import "../../../vendor/openzeppelin-solidity/v4.7.3/contracts/utils/Address.sol";
 import "./ArbitrumCrossDomainForwarder.sol";
