+++ contracts/src/v0.8/l2ep/dev/arbitrum/ArbitrumSequencerUptimeFeed.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.4;
 
 import {AddressAliasHelper} from "../../../vendor/arb-bridge-eth/v0.8.0-custom/contracts/libraries/AddressAliasHelper.sol";
-import {ForwarderInterface} from "./../../interfaces/ForwarderInterface.sol";
+import {ForwarderInterface} from "./../interfaces/ForwarderInterface.sol";
 import {AggregatorInterface} from "../../../interfaces/AggregatorInterface.sol";
 import {AggregatorV3Interface} from "../../../interfaces/AggregatorV3Interface.sol";
 import {AggregatorV2V3Interface} from "../../../interfaces/AggregatorV2V3Interface.sol";
 import {TypeAndVersionInterface} from "../../../interfaces/TypeAndVersionInterface.sol";
-import {FlagsInterface} from "./../../interfaces/FlagsInterface.sol";
-import {ArbitrumSequencerUptimeFeedInterface} from "../../interfaces/ArbitrumSequencerUptimeFeedInterface.sol";
+import {FlagsInterface} from "./../interfaces/FlagsInterface.sol";
+import {ArbitrumSequencerUptimeFeedInterface} from "../interfaces/ArbitrumSequencerUptimeFeedInterface.sol";
 import {SimpleReadAccessController} from "../../../SimpleReadAccessController.sol";
 import {ConfirmedOwner} from "../../../shared/access/ConfirmedOwner.sol";
