+++ contracts/src/v0.8/l2ep/dev/arbitrum/ArbitrumValidator.sol
@@ -8,9 +8,9 @@ import "../../../interfaces/AggregatorV3Interface.sol";
 import "../../../SimpleWriteAccessController.sol";
 
 /* ./dev dependencies - to be moved from ./dev after audit */
-import "../../interfaces/ArbitrumSequencerUptimeFeedInterface.sol";
-import "../../interfaces/FlagsInterface.sol";
-import "../../interfaces/IArbitrumDelayedInbox.sol";
+import "../interfaces/ArbitrumSequencerUptimeFeedInterface.sol";
+import "../interfaces/FlagsInterface.sol";
+import "../interfaces/IArbitrumDelayedInbox.sol";
 import "../../../vendor/arb-bridge-eth/v0.8.0-custom/contracts/libraries/AddressAliasHelper.sol";
 import "../../../vendor/@arbitrum/nitro-contracts/src/precompiles/ArbSys.sol";
 import "../../../vendor/openzeppelin-solidity/v4.7.3/contracts/utils/Address.sol";
