+++ contracts/src/v0.8/l2ep/dev/interfaces/ArbitrumSequencerUptimeFeedInterface.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+interface ArbitrumSequencerUptimeFeedInterface {
+  function updateStatus(bool status, uint64 timestamp) external;
+}
