+++ contracts/src/v0.8/l2ep/dev/interfaces/DelegateForwarderInterface.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+/// @title DelegateForwarderInterface - forwards a delegatecall to a target, under some conditions
+interface DelegateForwarderInterface {
+  /**
+   * @notice forward delegatecalls the `target` with `data`
+   * @param target contract address to be delegatecalled
+   * @param data to send to target contract
+   */
+  function forwardDelegate(address target, bytes memory data) external;
+}
