+++ contracts/src/v0.8/l2ep/dev/interfaces/ForwarderInterface.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+/// @title ForwarderInterface - forwards a call to a target, under some conditions
+interface ForwarderInterface {
+  /**
+   * @notice forward calls the `target` with `data`
+   * @param target contract address to be called
+   * @param data to send to target contract
+   */
+  function forward(address target, bytes memory data) external;
+}
