+++ contracts/test/v0.8/dev/ArbitrumValidator.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@ethereum-waffle/mock-contract'
 /// Pick ABIs from compilation
 // @ts-ignore
-import { abi as arbitrumSequencerStatusRecorderAbi } from '../../../artifacts/src/v0.8/dev/l2ep/arbitrum/ArbitrumSequencerUptimeFeed.sol/ArbitrumSequencerUptimeFeed.json'
+import { abi as arbitrumSequencerStatusRecorderAbi } from '../../../artifacts/src/v0.8/l2ep/dev/arbitrum/ArbitrumSequencerUptimeFeed.sol/ArbitrumSequencerUptimeFeed.json'
 // @ts-ignore
 import { abi as arbitrumInboxAbi } from '../../../artifacts/src/v0.8/vendor/arb-bridge-eth/v0.8.0-custom/contracts/bridge/interfaces/IInbox.sol/IInbox.json'
 // @ts-ignore
@@ -57,7 +57,7 @@ describe('ArbitrumValidator', () => {
 
     // Required for building the calldata
     arbitrumSequencerStatusRecorderFactory = await ethers.getContractFactory(
-      'src/v0.8/dev/l2ep/arbitrum/ArbitrumSequencerUptimeFeed.sol:ArbitrumSequencerUptimeFeed',
+      'src/v0.8/l2ep/dev/arbitrum/ArbitrumSequencerUptimeFeed.sol:ArbitrumSequencerUptimeFeed',
       deployer,
     )
     l1GasFeed = await deployMockContract(deployer as any, aggregatorAbi)
@@ -76,7 +76,7 @@ describe('ArbitrumValidator', () => {
 
     // Contract under test
     const arbitrumValidatorFactory = await ethers.getContractFactory(
-      'src/v0.8/dev/l2ep/arbitrum/ArbitrumValidator.sol:ArbitrumValidator',
+      'src/v0.8/l2ep/dev/arbitrum/ArbitrumValidator.sol:ArbitrumValidator',
       deployer,
     )
     arbitrumValidator = await arbitrumValidatorFactory.deploy(
