+++ contracts/test/v0.8/dev/OptimismCrossDomainForwarder.test.ts
@@ -26,7 +26,7 @@ before(async () => {
 
   // Contract factories
   forwarderFactory = await ethers.getContractFactory(
-    'src/v0.8/dev/l2ep/optimism/OptimismCrossDomainForwarder.sol:OptimismCrossDomainForwarder',
+    'src/v0.8/l2ep/dev/optimism/OptimismCrossDomainForwarder.sol:OptimismCrossDomainForwarder',
     owner,
   )
   greeterFactory = await ethers.getContractFactory(
